+++ kitchenpos-api/src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -8,8 +8,8 @@ import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.exception.NotFoundOrderTableException;
 import kitchenpos.table.exception.NotFoundTableGroupException;
-import kitchenpos.table.handler.TableGroupMapper;
-import kitchenpos.table.handler.TableGroupValidator;
+import kitchenpos.table.mapper.TableGroupMapper;
+import kitchenpos.table.validator.TableGroupValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
