+++ kitchenpos-api/src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -6,8 +6,8 @@ import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menu.exception.NotFoundMenuGroupException;
 import kitchenpos.menu.exception.NotFoundProductException;
-import kitchenpos.menu.handler.MenuMapper;
-import kitchenpos.menu.handler.MenuValidator;
+import kitchenpos.menu.mapper.MenuMapper;
+import kitchenpos.menu.validator.MenuValidator;
 import kitchenpos.wrap.Price;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
