+++ kitchenpos-api/src/test/java/kitchenpos/table/application/OrderTableServiceTest.java
@@ -14,8 +14,8 @@ import kitchenpos.table.exception.NotChangeNumberOfGuestThatEmptyTable;
 import kitchenpos.table.exception.NotChangeToEmptyThatCookingOrMealTable;
 import kitchenpos.table.exception.NotChangeToEmptyThatGroupTable;
 import kitchenpos.table.exception.NotFoundOrderTableException;
-import kitchenpos.table.handler.OrderTableMapper;
-import kitchenpos.table.handler.OrderTableValidator;
+import kitchenpos.table.mapper.OrderTableMapper;
+import kitchenpos.table.validator.OrderTableValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
