+++ kitchenpos-api/src/test/java/kitchenpos/table/application/TableGroupServiceTest.java
@@ -12,8 +12,8 @@ import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.exception.NotChangeToEmptyThatCookingOrMealTable;
 import kitchenpos.table.exception.NotEmptyOrExistTableGroupException;
 import kitchenpos.table.exception.NotFoundOrderTableException;
-import kitchenpos.table.handler.TableGroupMapper;
-import kitchenpos.table.handler.TableGroupValidator;
+import kitchenpos.table.mapper.TableGroupMapper;
+import kitchenpos.table.validator.TableGroupValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
