+++ kitchenpos-common/src/main/java/kitchenpos/menu/validator/MenuValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.handler;
+package kitchenpos.menu.validator;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
